move rcf param class to rcf package
