+++ common/src/main/java/org/opensearch/ml/common/input/parameter/rcf/BatchRCFParams.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.ml.common.input.parameter.ad;
+package org.opensearch.ml.common.input.parameter.rcf;
 
 import lombok.Builder;
 import lombok.Data;
