+++ ml-algorithms/src/test/java/org/opensearch/ml/engine/algorithms/rcf/BatchRandomCutForestTest.java
@@ -15,7 +15,7 @@ import org.opensearch.ml.common.dataframe.ColumnType;
 import org.opensearch.ml.common.dataframe.DataFrame;
 import org.opensearch.ml.common.dataframe.DefaultDataFrame;
 import org.opensearch.ml.common.dataframe.Row;
-import org.opensearch.ml.common.input.parameter.ad.BatchRCFParams;
+import org.opensearch.ml.common.input.parameter.rcf.BatchRCFParams;
 import org.opensearch.ml.common.FunctionName;
 import org.opensearch.ml.common.output.MLPredictionOutput;
 import org.opensearch.ml.common.Model;
