+++ ml-algorithms/src/test/java/org/opensearch/ml/engine/algorithms/rcf/FixedInTimeRandomCutForestTest.java
@@ -15,7 +15,7 @@ import org.opensearch.ml.common.dataframe.DefaultDataFrame;
 import org.opensearch.ml.common.FunctionName;
 import org.opensearch.ml.common.output.MLPredictionOutput;
 import org.opensearch.ml.common.Model;
-import org.opensearch.ml.common.input.parameter.ad.FitRCFParams;
+import org.opensearch.ml.common.input.parameter.rcf.FitRCFParams;
 
 import java.util.concurrent.ThreadLocalRandom;
